requirements/base.txt
# basics
Django==4.1.6
djangorestframework==3.14.0
django-extensions==3.2.1
drf-extra-fields=3.4.1
psycopg2==2.9.5
channels==4.0.0
daphne==4.0.0
substrate-interface==1.4.2
pillow==9.4.0
boto3==1.26.75